backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
Pillow==10.3.0
requests==2.32.3
reportlab==4.2.0
python-dotenv==1.0.1
gunicorn==22.0.0
